# go/root.go
/*
 * SWAPI
 *
 * This is a RESTful API SWAPI written in GO. You can find out more about SWAPI at [https://www.swapi.co](https://www.swapi.co). And you can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/). 
 *
 * API version: 1.0.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type Root struct {

	People string `json:"people,omitempty"`

	Planets string `json:"planets,omitempty"`

	Spaceships string `json:"spaceships,omitempty"`

	Vehicles string `json:"vehicles,omitempty"`

	Films string `json:"films,omitempty"`

	Species string `json:"species,omitempty"`
}